start going through to convert to causal length rather than delete wins
Still need to:
1. Process the `causal_length` value on non-causal length insertions to see if it requires resurrection

We do not need to process `causal_length` values on non-causal length insertions in order to kick off a delete given it would be impossible for these records to exist post deletion of a row. How could a user update a deleted row on their node in order to generate these records?